src/sprint1/requirements.txt
PyMySQL==1.0.2
python-dotenv==0.15.0
